# Faster BC1 quantization