I was investigating different methods of quantizing BC1 endpoints to R5G6B5 in hopes of improving quality. While I was unsuccessful in finding a higher-quality approach, I did find an approach of virtually equal quality that is faster.

## Background

BC1 blocks store 2 R5G6B5 colors as endpoints and then interpolate between these endpoints. This means that finding the best possible endpoints is a discrete optimization problem. However, since the continuous version of this problem is a lot easier to solve, the BC1 encoder first finds 2 endpoints in continuous space (aka f32) and then quantizes them. Therefore, how quantization is performed affects both quality and performance.

Simply rounding to the nearest R5G6B5 color is very fast, but produces suboptimal results.

To fix this, I previously optimized endpoints independently. The idea was to define an error metric that given 2 quantized endpoints measures the error between the best possible block using those endpoints and the input image to encode. For each channel $c$ of each endpoint, the algorithm measured the error for candidates with $\lfloor c \rfloor$ and $\lceil c \rceil$ for all channels. Since there are 3 channels, this gives $2^3=8$ quantized candidates per endpoint (because $\lfloor c \rfloor \ne \lceil c \rceil$ is typically the case). Instead of measuring all $8 \cdot 8 = 64$ possible combinations between the 2 endpoints, the endpoints optimized independently of each other. So only $8 \cdot 2=16$ combinations need to be measured. However, one combination for the second endpoint would be measured again, which means that we can skip it. So only 15 combinations need to be tested total.

Due to optimizing endpoints independently, I call this approach **endpoint-wise optimization**. This approach produces good results, but is quite expensive. The error metric is an expensive function to call, so reducing the number of tests is paramount for performance.

## The new approach

Instead of optimizing endpoints independently, it's possible to optimize both endpoints but one channel at a time. So hold all but 1 channel fixed and test the 4 possible combinations of floor and ceil for that channel in the 2 endpoints. This gives a total of $4*3=12$ candidates. However, only 10 need to be tested, since 2 overlap with the previous channels (similar to before).

I call this approach **channel-wise optimization**. Despite measuring 37% fewer candidates, this approach achieves virtually identical quality (see `test-data/encode_quality.md`). This reduces the time needed to encode BC1 images by around 4\~5% without loss of quality. (It's only 4\~5%, because quantization is only a part of the encoding process.)

## Optimized channel-wise

For the `Fast` quality setting, I wanted to have even faster quantization. The main idea was that if a channel $c$ is very close to $\lfloor c \rfloor$, then the optimal quantization of that channel is likely to be $\lfloor c \rfloor$. So if $c - \lfloor c \rfloor < r$ for some small threshold $r$, then don't bother checking candidates for $\lceil c \rceil$. Similar for $\lceil c \rceil$.

Using this idea, it is possible to cull unlikely candidates. The aggressiveness of the culling can be controlled using the threshold $r$, which allows for a smooth tradeoff between quality and performance. 

- $r=0$ means no culling and behaves the same as the normal channel-wise approach. 
- $r=0.5$ will cull all candidates and return the same result as simply rounding all channels. 

For now, I hard-coded it to $r=0.25$, which will result in 4.75 candidates being checked on average. This will worsen PSNR by -0.05\~-0.2dB, but also almost halves the total encoding time of `Fast`. IMO, it's a reasonable compromise between quality and performance.

## Benchmark

| Quality | Old      | New       | Change   |
| ------- | -------- | --------- | -------- |
| `Fast`    | 1.6882 ms | 912.21 µs | -46.064% |
| `Normal`  | 5.0254 ms | 4.7485 ms  | -5.5113% |
| `High`    | 13.817 ms | 13.304 ms  | -3.7131% |

Timings are measured using criterion, with uniform error, and on a single thread (`parallel: false`).

Especially for `Fast`, the speed-up is quite significant. It's now fast enough to encode a 4k image in under 1 second on a single thread.